+++ HARDWARE/control/control.h
@@ -7,12 +7,12 @@
 
 typedef struct
 {
-	float target_val;   // 目标值
-	float err;          // 偏差值
-	float err_last;     // 上一个偏差值
-	float Kp,Ki,Kd;     // 比例、积分、微分系数
-	float integral;     // 积分值
-	float output_val;   // 输出值
+	int target_val;   // 目标值
+	int err;          // 偏差值
+	int err_last;     // 上一个偏差值
+	int Kp,Ki,Kd;     // 比例、积分、微分系数
+	int integral;     // 积分值
+	int output_val;   // 输出值
 }PID;
 
 void moter_control(void);
